fix: refactor card and use card in FAQ styling in about us
